Squashed commit of the following:

commit 6eecedd09bf5aff5bec136878f90a8b3155731f7
Author: Russ Combs <rucombs@cisco.com>
Date:   Fri May 12 18:47:53 2017 -0400

    cleanup: fix misc formatting

commit 9db0cff83f841855d75670728c16c2b163af55c0
Author: Russ Combs <rucombs@cisco.com>
Date:   Fri May 12 18:28:56 2017 -0400

    cleanup: remove tedious and redundant option/flag comments

commit 4cc0ac76c7e965cd9f2f5dfe398efd2731536836
Author: Russ Combs <rucombs@cisco.com>
Date:   Fri May 12 18:20:45 2017 -0400

    alert_fast: lookup buffer ids when constructed

commit b48b39cf0bf4da3e1842c2f988508ba1b0160dd0
Author: Russ Combs <rucombs@cisco.com>
Date:   Fri May 12 14:10:36 2017 -0400

    expanded cmg logging for http buffers

commit 0785439d2dad33b718083db66911a3e628567a62
Author: Russ Combs <rucombs@cisco.com>
Date:   Sat May 6 17:17:52 2017 -0400

    cleanup: remove cruft

commit 669669fa43883c872536d1a472d43bc1d6298229
Author: Russ Combs <rucombs@cisco.com>
Date:   Sat May 6 16:43:03 2017 -0400

    log: print inspector::buffer[size]: on separate line

commit a0768e3c6fea1f76823c6814b0946f11b4376ef5
Author: Russ Combs <rucombs@cisco.com>
Date:   Fri May 5 14:28:19 2017 -0400

    make wide hex dump configurable

commit 9f0224dfd98ea6550cd4d5744c39ec6f0bb99a35
Author: Russ Combs <rucombs@cisco.com>
Date:   Wed May 3 14:53:09 2017 -0400

    revamp hex buffer dump format used by -A cmg
    -- add extra space at halfway points (both hex and ascii)

+++ src/loggers/alert_syslog.cc
@@ -202,10 +202,7 @@ static void AlertSyslog(
         if (event != NULL)
         {
             SnortSnprintfAppend(event_string, sizeof(event_string),
-                "[%lu:%lu:%lu] ",
-                (unsigned long)event->sig_info->gid,
-                (unsigned long)event->sig_info->sid,
-                (unsigned long)event->sig_info->rev);
+                "[%u:%u:%u] ", event->sig_info->gid, event->sig_info->sid, event->sig_info->rev);
         }
 
         if (msg != NULL)
